WIP: vm perf improvements

+++ chain/actors/actor_multisig.go
@@ -24,7 +24,7 @@ func init() {
 type MultiSigActor struct{}
 type MultiSigActorState struct {
 	Signers  []address.Address
-	Required uint32
+	Required uint64
 	NextTxID uint64
 
 	//TODO: make this map/sharray/whatever
@@ -59,7 +59,7 @@ type MTransaction struct {
 	Approved []address.Address
 	Complete bool
 	Canceled bool
-	RetCode  uint8
+	RetCode  uint64
 }
 
 func (tx MTransaction) Active() ActorError {
@@ -102,7 +102,7 @@ func (msa MultiSigActor) Exports() []interface{} {
 
 type MultiSigConstructorParams struct {
 	Signers  []address.Address
-	Required uint32
+	Required uint64
 }
 
 func (MultiSigActor) MultiSigConstructor(act *types.Actor, vmctx types.VMContext,
@@ -194,7 +194,7 @@ func (msa MultiSigActor) Propose(act *types.Actor, vmctx types.VMContext,
 		if aerrors.IsFatal(err) {
 			return nil, err
 		}
-		tx.RetCode = aerrors.RetCode(err)
+		tx.RetCode = uint64(aerrors.RetCode(err))
 		tx.Complete = true
 	}
 
@@ -229,12 +229,12 @@ func (msa MultiSigActor) Approve(act *types.Actor, vmctx types.VMContext,
 		}
 	}
 	tx.Approved = append(tx.Approved, vmctx.Message().From)
-	if uint32(len(tx.Approved)) >= self.Required {
+	if uint64(len(tx.Approved)) >= self.Required {
 		_, err := vmctx.Send(tx.To, tx.Method, tx.Value, tx.Params)
 		if aerrors.IsFatal(err) {
 			return nil, err
 		}
-		tx.RetCode = aerrors.RetCode(err)
+		tx.RetCode = uint64(aerrors.RetCode(err))
 		tx.Complete = true
 	}
 
@@ -319,7 +319,7 @@ func (msa MultiSigActor) RemoveSigner(act *types.Actor, vmctx types.VMContext,
 			newSigners = append(newSigners, s)
 		}
 	}
-	if params.Decrease || uint32(len(self.Signers)-1) < self.Required {
+	if params.Decrease || uint64(len(self.Signers)-1) < self.Required {
 		self.Required = self.Required - 1
 	}
 
@@ -366,7 +366,7 @@ func (msa MultiSigActor) SwapSigner(act *types.Actor, vmctx types.VMContext,
 }
 
 type MultiSigChangeReqParams struct {
-	Req uint32
+	Req uint64
 }
 
 func (msa MultiSigActor) ChangeRequirement(act *types.Actor, vmctx types.VMContext,
